feat: free trial end point (#324)
* made changes to the free trial end point

* Delete freeTrial.py

Deleted free trial page

* added .db to gitignore

* made changes
