+++ backend/app/.gitignore
@@ -143,4 +143,5 @@ cython_debug/
 #to allow frontend and backend see which code might break with github workflow
 
 .circleci
+
 *.db
